+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Entities/RobaUMagacinu.cs
@@ -48,7 +48,7 @@ namespace TD.Komercijalno.Contracts.Entities
         public double? WmsStanje { get; set; }
 
         [NotMapped]
-        public int Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public long Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         [NotMapped]
         public bool IsActive { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         [NotMapped]
