+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Entities/VrstaDokMag.cs
@@ -16,7 +16,7 @@ namespace TD.Komercijalno.Contracts.Entities
         public int MagacinId { get; set; }
 
         [NotMapped]
-        public int Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public long Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         [NotMapped]
         public bool IsActive { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         [NotMapped]
